requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
numpy==1.24.3
pillow==10.0.1
flask==2.2.5
imutils==0.5.4
requests==2.31.0
